backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
motor==3.5.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20